Document that FT_Get_X11_Font_Format() is a general call.
This function was originally added to return the underlying font
format in a way which was compatible with the needs of the XFree86
font server.  However, a function of this sort is also necessary when
freetype is used by PDF interpreters which apply different rules
based on the font format.

The X11 namespace is confusing for such users.  This commit ameliorates
that confusion by documenting that this function is in fact entirely
internal to freetype and doesn't depend on X11 in any way.