[builds] Add checks and targets for `docwriter'.
Add checks for Python, pip and docwriter. Docwriter is now a PyPI
package, downloadable from

  https://pypi.org/project/docwriter/

This package can be installed with the command `pip install docwriter'
on any system that has pip.

Running `make' on a fresh build will now warn if package `docwriter' is
not installed, and also provide information about the `make refdoc-venv'
alternative.

* builds/freetype.mk: Modify targats `refdoc' and `refdoc-venv' to
use detected python version/alias, and run docwriter as a package.

* builds/unix/configure.raw: Add rules to check python, pip and
docwriter. (code discussed in the dev mailing list).

* builds/unix/unix-def.in: Add variables PYTHON and PIP detected by
`configure.ac'. This is conditionally assigned (if empty) in
`freetype.mk' to the defaults.